+++ Source/Graphics/Pipeline/VkDiffuseGpuPipeline.cpp
@@ -57,8 +57,8 @@ namespace Injector
 		auto viewport = vk::Viewport(
 			0.0f,
 			0.0f,
-			surfaceExtent.width,
-			surfaceExtent.height,
+            static_cast<float>(surfaceExtent.width),
+            static_cast<float>(surfaceExtent.height),
 			0.0f,
 			1.0f);
 		auto scissor = vk::Rect2D(
@@ -199,8 +199,8 @@ namespace Injector
 			descriptorSetLayout);
 		auto descriptorSetAllocateInfo = vk::DescriptorSetAllocateInfo(
 			descriptorPool,
-			imageCount,
-			descriptorSetLayouts.data());
+            static_cast<uint32_t>(imageCount),
+            descriptorSetLayouts.data());
 
 		auto descriptorSets = std::vector<vk::DescriptorSet>(imageCount);
